Removed colored statusbar on KitKat as I could not resolve settings issues (may create custom Settins from blank Fragment and add colored statusbar back in later). Fixed some issues with tags

+++ App/src/main/java/com/caseybrooks/scripturememory/nowcards/votd/VOTD.java
@@ -155,9 +155,11 @@ public class VOTD {
                             && today.get(Calendar.DAY_OF_YEAR) == current.get(Calendar.DAY_OF_YEAR));
 
             if(!isCurrent) {
-                db.open();
-                db.deleteVerse(currentVerse);
-                db.close();
+				if(currentVerse.getMetadata().getInt(DefaultMetaData.STATE) != VerseDB.VOTD) {
+					db.open();
+					db.deleteVerse(currentVerse);
+					db.close();
+				}
                 currentVerse = null;
             }
         }
